requirements.txt
numpy
scipy
pillow
matplotlib
h5py
torchfile
scikit-learn
tqdm
word2vec
nltk